Add missing currency symbols to dictionary
I saw in a streamer's video that the peso comes up as PHP, but on my client at least it appears as ₱. That might be the work of their OBS plugin or it might be region dependent. I haven't checked. It doesn't matter either way since the additional symbols won't break anything. I just found it curious.